+++ package.json
@@ -22,6 +22,7 @@
   "type": "module",
   "devDependencies": {
     "ts-node": "^10.9.2",
+    "tsc-alias": "^1.8.13",
     "typescript": "^5.8.2"
   },
   "packageManager": "pnpm@9.14.2"
