+++ packages/core/package.json
@@ -17,7 +17,7 @@
     "LICENSE"
   ],
   "scripts": {
-    "build": "tsc",
+    "build": "tsc && tsc-alias",
     "npm-publish": "npm publish"
   },
   "repository": {
