+++ packages/core/src/index.ts
@@ -1,7 +1,5 @@
-export { createStore } from "./store";
-export { createForest } from "./forest";
-export { tree } from "./tree";
+export { createForest, tree } from "./dom";
+export * from "./router";
+export * from "./store";
+export * from "./types";
 export * from "./utilities";
-export * from "./trigger";
-export { router, addRoute } from "./router";
-export type * from "./types";
