+++ packages/forest-pure-demo/src/app/router/index.ts
@@ -1,0 +1,4 @@
+import { createRouter } from "@forest-js/core";
+
+// 라우터 인스턴스 생성 및 전역 관리
+export const { route, router } = createRouter();
