+++ packages/forest-pure-demo/src/pages/404/index.ts
@@ -1,0 +1,5 @@
+import { tree } from "@forest-js/core";
+
+const NotFound = tree("main");
+
+export default NotFound;
