+++ packages/forest-pure-demo/src/pages/about/index.ts
@@ -1,0 +1,5 @@
+import { tree } from "@forest-js/core";
+
+const About = tree("main");
+
+export default About;
