+++ packages/forest-pure-demo/src/pages/home/index.ts
@@ -1,0 +1,5 @@
+import { addChild, tree } from "@forest-js/core";
+
+const Home = addChild(["home", "home-title"])(tree("main"));
+
+export default Home;
